+++ examples/tutorial/lengthdemo.c
@@ -41,7 +41,7 @@ static void per_packet(libtrace_packet_t *packet)
 	 * Note that I use a while loop here to ensure that we correctly deal
 	 * with periods in which no packets are observed.
 	 */
-	while (ts.tv_sec > next_report) {
+	while ((uint32_t)ts.tv_sec > next_report) {
 
 		/* Print the timestamp for the report */
 		printf("%u\t", next_report);
